중복카드 점수보너스 업데이트

+++ Sum/Assets/NewScript/GameTimer.cs
@@ -133,7 +133,7 @@ public class GameTimer : MonoBehaviour
         }
     }
 
-    public void AddTime(float seconds)
+    public void AddTime(int seconds)
     {
         timeLeft += seconds;
     }
